feat: day 2 - integration boogaloo
